requirements.txt
flask
pyyaml
mailinglogger